requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

gradio==4.31.5
huggingface_hub==0.23.5

langgraph==0.0.55
langchain-core==0.2.43

requests==2.31.0
Pillow==10.2.0

neo4j==5.21.0
pinecone-client==3.2.2

python-dotenv==1.0.1

# Feature extraction service (ResNet50)

torch==2.3.1
torchvision==0.18.1

numpy==1.26.4
python-multipart==0.0.9